+++ Changelog.md
@@ -1,5 +1,6 @@
 ## v0.13.1 (unreleased)
 
+   * Fixed audio volume being reset on load every time
    * Fixed player escorts trying to respond to distress signals against the
      player
    * Fixed several missions using inexistent audio.soundPlay function